created Service directory and ResultService class
